Split topdown equality into separate file
